+++ src/components/InputNumber/constants.ts
@@ -1,0 +1,6 @@
+export const InputNumberMode = {
+  CURRENCY: 'currency',
+  DECIMAL: 'decimal'
+} as const;
+export type InputNumberMode =
+  (typeof InputNumberMode)[keyof typeof InputNumberMode];